+++ tests/test_pymagicc.py
@@ -106,7 +106,12 @@ def test_run_rcp85(package):
 
 @pytest.mark.slow
 def test_parameters(package):
-    results = run(rcp26, magicc_version=package.version, out_parameters=True, core_climatesensitivity=1.5)
+    results = run(
+        rcp26,
+        magicc_version=package.version,
+        out_parameters=True,
+        core_climatesensitivity=1.5,
+    )
     assert results.metadata["parameters"]["allcfgs"]["core_climatesensitivity"] == 1.5
     # Test removal of newlines in PARAMETERS.out
     assert "H\nFC134a" not in results.metadata["parameters"]["allcfgs"]["fgas_names"]
@@ -121,7 +126,13 @@ def test_default_config():
 
 @pytest.mark.slow
 def test_set_years(package):
-    results = run(rcp26, magicc_version=package.version, out_parameters=True, startyear=1900, endyear=2000)
+    results = run(
+        rcp26,
+        magicc_version=package.version,
+        out_parameters=True,
+        startyear=1900,
+        endyear=2000,
+    )
     assert results.metadata["parameters"]["years"]["startyear"] == 1900
     assert results.metadata["parameters"]["years"]["endyear"] == 2000
     assert results["time"].min().year == 1900